# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sinema API",
        "description": "Film, tür, kişi ve oyuncu yönetimi için RESTful API",
        "contact": {
            "name": "API Geliştirici",
            "email": "developer@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
        "/filmler": {
            "get": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Tüm filmleri listele",
                "description": "Veritabanındaki tüm filmleri tür ve oyuncu bilgileriyle birlikte getirir",
                "operationId": "1e1d05a9b1f9ec5a0851deef90507e85",
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Film"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Yeni film ekle",
                "description": "Veritabanına yeni bir film ekler",
                "operationId": "3c665a18e9bd6c091b1d4e93e132baf8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adi",
                                    "konusu",
                                    "tur_id"
                                ],
                                "properties": {
                                    "adi": {
                                        "description": "Film adı",
                                        "type": "string",
                                        "example": "The Matrix"
                                    },
                                    "konusu": {
                                        "description": "Film konusu",
                                        "type": "string",
                                        "example": "Bir bilgisayar programcısı..."
                                    },
                                    "imdb_puani": {
                                        "description": "IMDB puanı (0-10 arası)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.7
                                    },
                                    "tur_id": {
                                        "description": "Tür ID'si",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Film başarıyla oluşturuldu",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Film"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            }
        },
        "/filmler/{id}": {
            "get": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Belirli filmi getir",
                "description": "ID'ye göre belirli bir filmi tür ve oyuncu bilgileriyle birlikte getirir",
                "operationId": "ba209f8bccd09c8d78e07b3fdec9cffd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Film"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film bulunamadı"
                    }
                }
            },
            "put": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Film güncelle",
                "description": "Belirli bir filmin bilgilerini günceller",
                "operationId": "74398edc8d7c8d6ccda901ece91f8f99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "adi",
                                    "konusu",
                                    "tur_id"
                                ],
                                "properties": {
                                    "adi": {
                                        "description": "Film adı",
                                        "type": "string",
                                        "example": "The Matrix"
                                    },
                                    "konusu": {
                                        "description": "Film konusu",
                                        "type": "string",
                                        "example": "Bir bilgisayar programcısı..."
                                    },
                                    "imdb_puani": {
                                        "description": "IMDB puanı (0-10 arası)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 8.7
                                    },
                                    "tur_id": {
                                        "description": "Tür ID'si",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Film başarıyla güncellendi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Film"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film bulunamadı"
                    },
                    "422": {
                        "description": "Doğrulama hatası"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Film sil",
                "description": "Belirli bir filmi veritabanından siler",
                "operationId": "5b1dfe8101c7213abd2dbc9bfb234268",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Film ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Film başarıyla silindi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Film başarıyla silindi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Film bulunamadı"
                    }
                }
            }
        },
        "/filmler/tur/{tur_id}": {
            "get": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Türe göre filmleri listele",
                "description": "Belirli bir türe ait filmleri getirir",
                "operationId": "29c5f7b6397d70adbde65d480c959ee8",
                "parameters": [
                    {
                        "name": "tur_id",
                        "in": "path",
                        "description": "Tür ID'si",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Film"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/filmler/search/{query}": {
            "get": {
                "tags": [
                    "Filmler"
                ],
                "summary": "Film ara",
                "description": "Film adı veya konusuna göre arama yapar",
                "operationId": "1d159e8b2154e1df34a07b736deff15c",
                "parameters": [
                    {
                        "name": "query",
                        "in": "path",
                        "description": "Arama terimi",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Matrix"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Başarılı",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Film"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Film": {
                "title": "Film",
                "description": "Film modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Film adı",
                        "type": "string",
                        "example": "The Matrix"
                    },
                    "konusu": {
                        "description": "Film konusu",
                        "type": "string",
                        "example": "Bir bilgisayar programcısı..."
                    },
                    "imdb_puani": {
                        "description": "IMDB puanı",
                        "type": "number",
                        "format": "float",
                        "example": 8.7
                    },
                    "tur_id": {
                        "description": "Tür ID'si",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "tur": {
                        "$ref": "#/components/schemas/Tur"
                    },
                    "oyuncular": {
                        "description": "Film oyuncuları",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Oyuncu"
                        }
                    }
                },
                "type": "object"
            },
            "Kisi": {
                "title": "Kişi",
                "description": "Kişi modeli (oyuncu, yönetmen vs.)",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Kişi adı",
                        "type": "string",
                        "example": "Keanu Reeves"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Oyuncu": {
                "title": "Oyuncu",
                "description": "Film-kişi ilişki modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "film_id": {
                        "description": "Film ID'si",
                        "type": "integer",
                        "example": 1
                    },
                    "kisi_id": {
                        "description": "Kişi ID'si",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "film": {
                        "$ref": "#/components/schemas/Film"
                    },
                    "kisi": {
                        "$ref": "#/components/schemas/Kisi"
                    }
                },
                "type": "object"
            },
            "Tur": {
                "title": "Tür",
                "description": "Film türü modeli",
                "properties": {
                    "id": {
                        "description": "Benzersiz ID",
                        "type": "integer",
                        "example": 1
                    },
                    "adi": {
                        "description": "Tür adı",
                        "type": "string",
                        "example": "Aksiyon"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-26T20:27:34.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-26T20:27:34.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}